feat: detail screen for testing
